LayoutTests/http/wpt/service-workers/media-range-request-expected.txt

CONSOLE MESSAGE: Media response origin validation failed.
CONSOLE MESSAGE: Media response origin validation failed.


PASS  Validate range response rewriting with regular fetch
PASS  Validate range response rewriting with synthetic response
